updating import path for moved files
