+++ dal/backup/backup.py
@@ -15,7 +15,7 @@ from datetime import datetime
 from zipfile import ZipFile
 
 from movai_core_shared.logger import Log
-from dal.scopes.scopestree import ScopesTree, scopes
+from dal.models.scopestree import ScopesTree, scopes
 from dal.models.model import Model
