+++ dal/classes/utils/acl.py
@@ -15,7 +15,7 @@ from miracle import Acl
 from movai_core_shared.envvars import REST_SCOPES
 from movai_core_shared.logger import Log
 
-from dal.scopes.scopestree import ScopesTree, scopes
+from dal.models.scopestree import ScopesTree, scopes
 
 
 class ACLManager:
