+++ dal/helpers/parsers.py
@@ -11,7 +11,7 @@ import ast
 import re
 import os
 from movai_core_shared.logger import Log
-from dal.scopes.scopestree import scopes
+from dal.models.scopestree import scopes
 from dal.models.var import Var
