+++ dal/models/baseuser.py
@@ -17,7 +17,7 @@ from movai_core_shared.exceptions import (
     UserPermissionsError,
 )
 
-from dal.scopes.scopestree import ScopesTree, scopes
+from dal.models.scopestree import ScopesTree, scopes
 from dal.models.model import Model
 from dal.classes.utils.acl import NewACLManager
