+++ dal/models/ldapconfig.py
@@ -7,7 +7,7 @@ from movai_core_shared.exceptions import (
     LdapConfigMissingParameter)
 
 from dal.models.model import Model
-from dal.scopes.scopestree import ScopesTree, scopes
+from dal.models.scopestree import ScopesTree, scopes
