+++ dal/models/node.py
@@ -9,7 +9,7 @@
 """
 from movai_core_shared.consts import ROS1_NODELET, MOVAI_STATE, ROS1_PLUGIN
 
-from dal.scopes.scopestree import scopes
+from dal.models.scopestree import scopes
 from dal.models.model import Model
